added filtrating and paginating

+++ Reputy.Contracts/Advertisement/AdvertisementResponse.cs
@@ -5,7 +5,7 @@
         public Guid ID { get; init; }
         public string? ImageUrl { get; init; }
         public string Title { get; init; }
-        public decimal Price { get; init; }
+        public int Price { get; init; }
         public AdvertisementRealEstateResponse AdvertisementRealEstate { get; init; }
     }
